\name{add_nav}
\alias{add_nav}
\title{Add a navigation marker}
\usage{
  add_nav(title, subhead, header = TRUE)
}
\arguments{
  \item{title}{Name for link and possible header}

  \item{subhead}{when \code{header=TRUE} add a sub heading
  in small}

  \item{header}{if \code{TRUE} add a header to the page in
  addition to the anchor.}
}
\description{
  Adds a link in the top nav bar and an anchor in the file.
  Also can add a header to the page.
}

