\name{alert}
\alias{alert}
\title{Make a dismissable label}
\usage{
  alert(title = "", txt,
    type = c("block", "error", "success", "info"))
}
\arguments{
  \item{title}{title of label}

  \item{txt}{HTML formatted text}

  \item{type}{type of label}
}
\description{
  Make a dismissable label
}

