\name{badge}
\alias{badge}
\title{Make a badge}
\usage{
  badge(txt,
    type = c("success", "warning", "important", "info", "inverse"),
    id)
}
\arguments{
  \item{txt}{the text, typically a number}

  \item{type}{empty for the default, else specifies the
  class}

  \item{id}{optional unique DOM id for object if wanting to
  adjust later}
}
\description{
  Badges are typically numbered
}

