\name{call_ref_meth}
\alias{call_ref_meth}
\title{Call a method}
\usage{
  call_ref_meth(obj, meth, params, request)
}
\arguments{
  \item{obj}{object (Reference class)}

  \item{meth}{method name}

  \item{params}{list of argument}

  \item{request}{httpd request object passed into
  environment of call}
}
\description{
  Call a method
}

