% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{children}
\alias{children}
\title{A fertility survey - "children" table}
\format{
a data frame containing one record for each child of the surveyed women in the \link{fertility} survey.
}
\usage{
children
}
\description{
Some fictive results from a fecondity survey.
}
\keyword{datasets}
