% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{enfants}
\alias{enfants}
\title{A fertility survey - "enfants" table}
\format{a data frame containing one record for each child of the surveyed women in the \link{fecondite} survey.}
\usage{
enfants
}
\description{
Some fictive results from a fecondity survey.
}
\keyword{datasets}
