% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.multi.R
\name{escape_regex}
\alias{escape_regex}
\title{Escape regex special chars
Code directly taken from Hmisc::escapeRegex}
\usage{
escape_regex(s)
}
\description{
Escape regex special chars
Code directly taken from Hmisc::escapeRegex
}

