% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{femmes}
\alias{femmes}
\title{A fertility survey - "femmes" table}
\format{
a data frame containing the questionnaire administered to all 15-49 years old women 
       living in the selected households for the \link{fecondite} survey.
}
\description{
Some fictive results from a fecondity survey.
}
\keyword{datasets}
