% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{fertility}
\alias{fertility}
\title{A fertility survey}
\format{
3 data frames with labelled data (as if data would have been imported from SPSS with \pkg{haven}):
   \itemize{
     \item \code{households} contains some information from the households selected for the survey;
     \item \code{women} contains the questionnaire administered to all 15-49 years old women
       living in the selected households;
     \item\code{children} contains one record for each child of the surveyed women.
   }

   Data can be linked using the variables \code{id_household} and \code{id_woman}.
}
\description{
Some fictive results from a fecondity survey, with English labels.
}
\examples{
data(fertility)
describe(households)
describe(women)
describe(children)
}
\seealso{
\link{fecondite} for an French version of this dataset.
}
\keyword{datasets}
