% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-utils.R
\name{first_non_null}
\alias{\%||\%}
\alias{first_non_null}
\title{Return first non-null of two values}
\usage{
x \%||\% y
}
\arguments{
\item{x}{first object}

\item{y}{second object}
}
\description{
Return first non-null of two values
}

