\name{form_action}
\alias{form_action}
\title{Create javascript for processing a form using serializeArray.}
\usage{
  form_action(form_id, script_name, success)
}
\arguments{
  \item{form_id}{DOM id of form}

  \item{script_name}{script name after base url}
}
\description{
  The post require `get_post_from_raw(request)` to be read
  properly. After processing POST request pass back url of
  next page.
}

