% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{households}
\alias{households}
\title{A fertility survey - "households" table}
\format{
a data frame containing some information from the households selected for the \link{fertility} survey.
}
\description{
Some fictive results from a fecondity survey.
}
\keyword{datasets}
