% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-utils.R
\name{ifunc_run_as_addin}
\alias{ifunc_run_as_addin}
\title{Check if we are currently running as an rstudio addin}
\usage{
ifunc_run_as_addin()
}
\description{
Check if we are currently running as an rstudio addin
}

