% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-utils.R
\name{ifunc_show_alert}
\alias{ifunc_show_alert}
\title{Display an alert, only on first launch for the current session}
\usage{
ifunc_show_alert(run_as_addin)
}
\arguments{
\item{run_as_addin}{TRUE if the function is running as an rstudio addin}
}
\description{
Display an alert, only on first launch for the current session
}
