\name{input_radio}
\alias{input_radio}
\title{a radio button input}
\usage{
  input_radio(id = NULL, values, labels = values,
    initial = -1, collapse = "", ...)
}
\arguments{
  \item{id}{DOM id}

  \item{values}{values passed back}

  \item{labels}{optional labels for values}

  \item{initial}{integer index or one or matched against
  the values}

  \item{collapse}{set to \code{"</br>"} to make vertical}
}
\description{
  a radio button input
}

