% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{menages}
\alias{menages}
\title{A fertility survey - "menages" table}
\format{a data frame containing some information from the households selected for the \link{fecondite} survey.}
\description{
Some fictive results from a fecondity survey.
}
\keyword{datasets}

