% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookfor.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{look_for}
\alias{lookfor}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{labelled}{\code{\link[labelled]{look_for}}, \code{\link[labelled:look_for]{lookfor}}}
}}

