% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{rp2012}
\alias{rp2012}
\title{2012 French Census - French cities of more than 2000 inhabitants}
\format{
A data frame with 5170 rows and 60 variables
}
\source{
\url{http://www.insee.fr/fr/bases-de-donnees/default.asp?page=recensements.htm}
}
\description{
Sample from the 2012 national french census. It contains results for every french city of more than 2000 inhabitants, and a small subset of variables, both in population counts and proportions.
}
\keyword{datasets}
