% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{rp2018}
\alias{rp2018}
\title{2018 French Census - French cities of more than 2000 inhabitants}
\format{
A data frame with 5417 rows and 62 variables
}
\source{
\url{https://www.insee.fr/fr/information/5369871}
}
\description{
Sample from the 2018 national french census. It contains results for every
french city of more than 2000 inhabitants, and a small subset of variables,
both in population counts and proportions.
}
\keyword{datasets}
