% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{rp99}
\alias{rp99}
\title{1999 French Census - Cities from the Rhône state}
\format{
A data frame with 301 rows and 21 variables
}
\source{
\url{https://www.insee.fr/fr/information/2008354}
}
\description{
Sample from the 1999 french census for the cities of the Rhône state.
}
\keyword{datasets}
