% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{women}
\alias{women}
\title{A fertility survey - "women" table}
\format{a data frame containing the questionnaire administered to all 15-49 years old women 
       living in the selected households for the \link{fertility} survey.}
\usage{
women
}
\description{
Some fictive results from a fecondity survey.
}
\keyword{datasets}
