% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighting.R
\name{Xtabs}
\alias{Xtabs}
\title{Weighted Crosstabs}
\usage{
Xtabs(df, x, y = NULL, type = "percent", weight = NULL, normwt = FALSE,
  na.rm = TRUE, na.show = FALSE, digits = 3, exclude = NULL)
}
\arguments{
\item{df}{A data.frame that contains \code{x} and (optionally) \code{y} and \code{weight}.}

\item{x}{variable name (found in \code{df}). Xtabs(my.data, x = 'q1').}

\item{y}{one (or more) variable names. Xtabs(my.data, x = 'q1', y = c('sex', 'race')).}

\item{type}{'percent', 'proportion', or 'counts' (type of table returned).}

\item{weight}{variable name for weight (found in \code{df}).}

\item{normwt}{if TRUE, normalize weights so that the total weighted count is the same as the unweighted one}

\item{na.rm}{if TRUE, remove NA values before computation}

\item{na.show}{if TRUE, show NA count in table output}

\item{digits}{Number of significant digits.}

\item{exclude}{values to remove from x and y. To exclude NA, use na.rm argument.}
}
\description{
Generate table with multiple weighted crosstabs (full sample is first column).
kable(), which is found in library(knitr), is recommended for use with RMarkdown.
}
\details{
Xtabs calls wtd.table on `\code{x}` and, as applicable, each variable named by `\code{y}`.
}
\examples{
data(hdv2003) 
Xtabs(hdv2003, x = "relig", y = c("qualif", "trav.imp"), weight = "poids")
Xtabs(hdv2003, x = "relig", y = c("qualif", "trav.imp"), type = "counts")
# library(knitr)
# xt <- Xtabs(hdv2003, x = "relig", y = c("qualif", "trav.imp"), weight = "poids")
# kable(xt)                        # to use with RMarkdown...

}
