% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.multi.R
\name{escape_regex}
\alias{escape_regex}
\title{Escape regex special chars
Code directly taken from Hmisc::escapeRegex}
\usage{
escape_regex(s)
}
\arguments{
\item{s}{string to escape regex special chars from}
}
\description{
Escape regex special chars
Code directly taken from Hmisc::escapeRegex
}
